pyproject.toml
[tool.poetry]
name = "venmo-auto-transfer"
version = "0.1.0"
description = ""
authors = ["Radian LLC <contact+venmo-auto-transfer@radian.codes>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
